Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I had created the ifdefs to check for specific versions at the time because I wasn't sure if this was going to be changed in >4.7.0.

At this point, it makes sense to expand the ifdefs to cover all of 4.7.x.

Patch-set: 2